+++ src/plone/app/blob/adapters/stringio.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from plone.app.blob.interfaces import IBlobbable
 from plone.app.blob.utils import guessMimetype
-from StringIO import StringIO
+from six import StringIO
 from zope.component import adapts
 from zope.interface import implementer