+++ src/plone/app/blob/field.py
@@ -27,11 +27,13 @@ from Products.Archetypes.atapi import PrimaryFieldMarshaller
 from Products.Archetypes.Registry import registerField
 from Products.Archetypes.utils import contentDispositionHeader
 from Products.CMFCore.permissions import View
-from StringIO import StringIO
+from six import StringIO
 from transaction import savepoint
 from webdav.common import rfc1123_date
 from ZODB.blob import Blob
 from zope.interface import implementer
+
+import six
 
 
 @implementer(IWebDavUpload)
@@ -91,8 +93,8 @@ class BlobWrapper(Implicit, Persistent):
 
         filename = self.getFilename()
         if filename is not None:
-            if not isinstance(filename, unicode):
-                filename = unicode(filename, charset, errors='ignore')
+            if not isinstance(filename, six.text_type):
+                filename = six.text_type(filename, charset, errors='ignore')
             filename = IUserPreferredFileNameNormalizer(
                 REQUEST
             ).normalize(
@@ -175,7 +177,7 @@ class BlobWrapper(Implicit, Persistent):
     @security.private
     def setFilename(self, value):
         """ set filename for this blob """
-        if isinstance(value, basestring):
+        if isinstance(value, six.string_types):
             value = value[max(value.rfind('/'),
                               value.rfind('\\'),
                               value.rfind(':')) + 1:]
@@ -240,14 +242,14 @@ class BlobField(ObjectField):
         # create a new blob instead of modifying the old one to
         # achieve copy-on-write semantics
         blob = BlobWrapper(self.default_content_type)
-        if isinstance(value, basestring):
+        if isinstance(value, six.string_types):
             value = StringIO(value)     # simple strings cannot be adapted...
             setattr(value, 'filename', kwargs.get('filename', None))
         if value is not None:
             blobbable = IBlobbable(value)
             try:
                 blobbable.feed(blob.getBlob())
-            except ReuseBlob, exception:
+            except ReuseBlob as exception:
                 blob.setBlob(exception.args[0])     # reuse the given blob
             mimetype = kwargs.get('mimetype', None)
             if not mimetype:
@@ -282,8 +284,8 @@ class BlobField(ObjectField):
                     request.form.get('title')
                 ):
                     return  # don't rename now if AT should do it from title
-            if not isinstance(filename, unicode):
-                filename = unicode(filename, instance.getCharset())
+            if not isinstance(filename, six.text_type):
+                filename = six.text_type(filename, instance.getCharset())
             filename = IUserPreferredFileNameNormalizer(
                 request
             ).normalize(